python-scraper/requirements.txt
# BizCompare Pro - Python Data Scraping System
# Collects real business data from government and commercial sources

# Core Dependencies
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.26.0
supabase>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
schedule>=1.2.0
tenacity>=8.2.0
pyyaml>=6.0.0
rich>=13.7.0
lxml>=4.9.0
apscheduler>=3.10.0
httpx>=0.25.0
